Fix all compiler warnings: unused vars, sign-compare, reorder, format
- shortcut.h: fix sign-compare (size_t vs qsizetype in loop)
- NppIO.h: add parentheses around && in || (warn)
- NppBigSwitch.cpp: fix %d format for size_t (format warning)
- Application.cpp: fix uninitialized warning (activeBuffer=nullptr),
  unused path/encoding params, unused variable r
- FileLoaderWorker: fix member init order (reorder warning)
- FileHelper: check mkstemp return value
- StringHelper: use chars param in indentLines
- ShortcutManager: remove unused _parseModifiers static function
- ScintillaEditView: remove unused nextLineEnd, delta vars
- DocTabView: add (void)w for unused ed cast
- ScintillaCtrls: remove unused len var
- TabBar: remove unused diff, closeAllAction vars, dpi var
- StatusBar: remove unused action vars in contextMenuEvent
- MainWindow: fix unused editMenu/searchMenu/etc vars,
  unused editor/buf vars
- FindReplaceDialog: remove unused foundEnd, lineCount vars
- GoToLineDialog: remove unused column var
- SmartHighlighter: remove unused searchStr, fix match vars

+++ src/MISC/Common/shortcut.h
@@ -131,7 +131,7 @@ public:
     void clearDups();
     friend inline bool operator==(const ScintillaKeyMap& a, const ScintillaKeyMap& b) {
         if (a._keyCombos.size() != b._keyCombos.size()) return false;
-        for (size_t i = 0; i < a._keyCombos.size(); ++i)
+        for (size_t i = 0; i < static_cast<size_t>(a._keyCombos.size()); ++i)
             if (a._keyCombos[i]._isCtrl != b._keyCombos[i]._isCtrl || a._keyCombos[i]._isAlt != b._keyCombos[i]._isAlt || a._keyCombos[i]._isShift != b._keyCombos[i]._isShift || a._keyCombos[i]._key != b._keyCombos[i]._key) return false;
         return true;
     }
